Revert "Fix MSVC /Wall /WX build for aie_codegen and apply BinSkim-oriented flags (#138)"
This reverts commit 5640a67844cf4e8a0ab1b89432eab7b74e39aac0.

+++ src/common/xaie_helper.c
@@ -1767,7 +1767,7 @@ void _XAie_ResetInUseBitMaps(XAie_DevInst *DevInst)
 	// Calculate the start bit position for each column and reset the bits in
 	// the global bitmaps.
 	for (int col = startCol; col < (startCol + numCols); col++) {
-		XAie_LocType ColStartLoc = { 0U, (u8)col };
+		XAie_LocType ColStartLoc = { 0, col };
 		u32 ColStartBitPos = _XAie_GetTileBitPosFromLoc(DevInst, ColStartLoc);
 		_XAie_ClrBitInBitmap(DevInst->DevOps->TilesInUse, ColStartBitPos, \
 			DevInst->NumRows);
