+++ src/device/xaie_device_aie2ps.c
@@ -246,7 +246,7 @@ AieRC _XAie2PS_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 
 	/*Enable the clock control register for shims*/
 	for(u32 C = Args->StartCol; C < (Args->StartCol + Args->NumCols); C++) {
-		XAie_LocType TileLoc = XAie_TileLoc((u8)C, 1U);
+		XAie_LocType TileLoc = XAie_TileLoc(C, 1);
 
 		RC = _XAie2PS_PmSetColumnClockBuffer(DevInst, TileLoc,
 				Args->Enable);
@@ -264,8 +264,7 @@ AieRC _XAie2PS_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 		}
 	}
 
-	TileStatus = _XAie_GetTileBitPosFromLoc(DevInst,
-			XAie_TileLoc((u8)Args->StartCol, 1U));
+	TileStatus = _XAie_GetTileBitPosFromLoc(DevInst, XAie_TileLoc(Args->StartCol, 1));
 	NumTiles =(u32)((DevInst->NumRows - 1U) * (Args->NumCols));
 
 	if(Args->Enable) {
